EditorFix
Merged My and MoonKill's version into an unholy hybrid. It is now
compatible with android, only uses the mouse. You can edit the cell,
place cells and destroy them. It now nicely zooms in. I learned how to
use enums and events, so there is a lot of that in this update