AI rework
Semi implemented new way to sort cells on change.
Added player script to hold player controlled cells for easier AI integration.
Target lists are now used to determine cell sorting on change.
but are always empty so the sorting fails.

Scenario when AI overtakes neutral cell results in no cells being added as targets to Ais that register the AI that overtook the cell as an Enemy
Others might appear with more testing.